mod) move configuration files into config dir
